+++ components/modals/EmergencyContactModal.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import BaseModal from "./BaseModal";
-import FormField from "../FormField";
+import FormField from "../Form/FormField";
 import CustomButton from "../Buttons/CustomButton";
 
 const EmergencyContactModal = ({
